refac: message timestamp
